chore(molecule) split vm agent healthy tests from log ones
